Combine deploy and package service files
