+++ packageService/index.js
@@ -1,6 +1,25 @@
+var zipFolder = new require('zip-folder');
+var fs = require('fs');
 var request = require('request-promise');
 
 module.exports = {
+  deploy: function(config, logger, location) {
+    logger.info('\nDeploying package');
+    return createZip(location).then(zip => {
+      return request.post({
+        uri: config.endpoint + '/packages',
+        qs: {
+          tenantId: config.tenantId
+        },
+        formData: {
+          file: fs.createReadStream(zip)
+        }
+      });
+    }).then((body) => {
+      logger.info('\nDone deploying');
+    });
+  },
+
   listInstalled: function(config, logger) {
     logger.debug('\nListing installed packages');
     return request({
@@ -18,3 +37,17 @@ module.exports = {
     });
   }
 };
+
+
+function createZip(location) {
+  var zip = './package.zip';
+  return new Promise((resolve, reject) => {
+    zipFolder(location, zip, function(err) {
+      if (err) {
+        reject(err);
+      } else {
+        resolve(zip);
+      }
+    });
+  });
+}
